requirements.txt
absl-py==0.7.0
astor==0.7.1
Click==7.0
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.0
decorator==4.3.2
Django==2.1.5
dlib==19.16.0
Flask==1.0.2
gast==0.2.2
grpcio==1.18.0
gunicorn==19.9.0
h5py==2.9.0
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
networkx==2.2
numpy==1.15.0
opencv-python==4.0.0.21
pandas==0.24.0
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.2
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.14.1
